Completion of Widgets, Apps Panel
When another app opens the Apps Panel, it returns the item tapped by the user in an appOut variable. This can be used to set default apps or for similar purposes. Additionally, there are some additions and finalizations of widgets. This implementation will soon extend to Watch Faces for additional customization.